refactor: improve test organization
