+++ test/integration/dispose/sync/pool-cluster.test.mts
@@ -1,0 +1,63 @@
+import type { PoolCluster, RowDataPacket } from '../../../../index.js';
+import { assert, describe, it, skip } from 'poku';
+import { createPoolCluster, getConfig } from '../../../common.test.mjs';
+
+if (!('dispose' in Symbol)) {
+  skip('Symbol.dispose is not supported in this runtime');
+}
+
+const clusterQuery = (cluster: PoolCluster, sql: string) =>
+  new Promise<RowDataPacket[]>((resolve, reject) => {
+    cluster.of('*').query<RowDataPacket[]>(sql, (err, rows) => {
+      if (err) return reject(err);
+      resolve(rows);
+    });
+  });
+
+await describe('PoolCluster should implement Symbol.dispose', async () => {
+  const cluster = createPoolCluster();
+
+  cluster.add('MASTER', getConfig());
+
+  it('should be a function', () => {
+    assert.strictEqual(typeof cluster[Symbol.dispose], 'function');
+  });
+
+  cluster[Symbol.dispose]();
+});
+
+await describe('dispose should end the pool cluster', async () => {
+  const cluster = createPoolCluster();
+
+  cluster.add('MASTER', getConfig());
+
+  const rows = await clusterQuery(cluster, 'SELECT 1');
+
+  cluster[Symbol.dispose]();
+
+  it('should have received the query result', () => {
+    assert.deepStrictEqual(rows, [{ 1: 1 }]);
+  });
+
+  it('should have closed the pool cluster', () => {
+    // @ts-expect-error: internal access
+    assert.strictEqual(cluster._closed, true);
+  });
+});
+
+await describe('dispose should handle end before dispose on pool cluster', async () => {
+  const cluster = createPoolCluster();
+
+  cluster.add('MASTER', getConfig());
+
+  await new Promise<void>((resolve, reject) => {
+    cluster.end((err) => (err ? reject(err) : resolve()));
+  });
+
+  cluster[Symbol.dispose]();
+
+  it('should have closed the pool cluster', () => {
+    // @ts-expect-error: internal access
+    assert.strictEqual(cluster._closed, true);
+  });
+});
