+++ test/integration/dispose/sync/pool.test.mts
@@ -1,17 +1,6 @@
-import type {
-  Connection,
-  Pool,
-  PoolCluster,
-  PoolConnection,
-  RowDataPacket,
-} from '../../../index.js';
+import type { Pool, PoolConnection, RowDataPacket } from '../../../../index.js';
 import { assert, describe, it, skip } from 'poku';
-import {
-  createConnection,
-  createPool,
-  createPoolCluster,
-  getConfig,
-} from '../../common.test.mjs';
+import { createPool } from '../../../common.test.mjs';
 
 if (!('dispose' in Symbol)) {
   skip('Symbol.dispose is not supported in this runtime');
@@ -25,7 +14,7 @@ const getConnection = (pool: Pool) =>
     });
   });
 
-const query = (conn: Connection, sql: string) =>
+const query = (conn: PoolConnection, sql: string) =>
   new Promise<RowDataPacket[]>((resolve, reject) => {
     conn.query<RowDataPacket[]>(sql, (err, rows) => {
       if (err) return reject(err);
@@ -50,8 +39,8 @@ await describe('using should release the connection back to the pool', async () 
 
   await it('should use and dispose the connection', async () => {
     using conn = await getConnection(pool);
-
     const rows = await query(conn, 'SELECT 1');
+
     assert.deepStrictEqual(rows, [{ 1: 1 }]);
   });
 
@@ -68,15 +57,15 @@ await describe('Pool should serve a new connection after using releases the prev
 
   await it('should use and dispose the first connection', async () => {
     using conn = await getConnection(pool);
-
     const rows = await query(conn, 'SELECT 1');
+
     assert.deepStrictEqual(rows, [{ 1: 1 }]);
   });
 
   await it('should use and dispose the second connection', async () => {
     using conn = await getConnection(pool);
-
     const rows = await query(conn, 'SELECT 1');
+
     assert.deepStrictEqual(rows, [{ 1: 1 }]);
   });
 
@@ -92,8 +81,12 @@ await describe('dispose should release the connection', async () => {
   const pool = createPool({ connectionLimit: 1 });
   const conn = await getConnection(pool);
   const rows = await query(conn, 'SELECT 1');
-  assert.deepStrictEqual(rows, [{ 1: 1 }]);
+
   conn[Symbol.dispose]();
+
+  it('should have received the query result', () => {
+    assert.deepStrictEqual(rows, [{ 1: 1 }]);
+  });
 
   it('should have returned the connection to the free pool', () => {
     // @ts-expect-error: internal access
@@ -108,10 +101,10 @@ await describe('using should handle manual `destroy` before automatic dispose', 
 
   await it('should not error when connection is destroyed within using', async () => {
     using conn = await getConnection(pool);
-
     const rows = await query(conn, 'SELECT 1');
-    assert.deepStrictEqual(rows, [{ 1: 1 }]);
+
     conn.destroy();
+    assert.deepStrictEqual(rows, [{ 1: 1 }]);
   });
 
   it('should have removed the destroyed connection from the pool', () => {
@@ -120,41 +113,6 @@ await describe('using should handle manual `destroy` before automatic dispose', 
   });
 
   await pool.promise().end();
-});
-
-await describe('Connection should implement Symbol.dispose', async () => {
-  const conn = createConnection();
-
-  it('should be a function', () => {
-    assert.strictEqual(typeof conn[Symbol.dispose], 'function');
-  });
-
-  conn.end();
-});
-
-await describe('dispose should end the connection', async () => {
-  const conn = createConnection();
-  const rows = await query(conn, 'SELECT 1');
-  assert.deepStrictEqual(rows, [{ 1: 1 }]);
-  conn[Symbol.dispose]();
-
-  it('should have closed the connection', () => {
-    // @ts-expect-error: internal access
-    assert.strictEqual(conn._closing, true);
-  });
-});
-
-await describe('dispose should handle destroy before dispose on connection', async () => {
-  const conn = createConnection();
-  const rows = await query(conn, 'SELECT 1');
-  assert.deepStrictEqual(rows, [{ 1: 1 }]);
-  conn.destroy();
-  conn[Symbol.dispose]();
-
-  it('should have closed the connection', () => {
-    // @ts-expect-error: internal access
-    assert.strictEqual(conn._closing, true);
-  });
 });
 
 await describe('Pool should implement Symbol.dispose', async () => {
@@ -171,9 +129,13 @@ await describe('dispose should end the pool', async () => {
   const pool = createPool({ connectionLimit: 1 });
   const conn = await getConnection(pool);
   const rows = await query(conn, 'SELECT 1');
-  assert.deepStrictEqual(rows, [{ 1: 1 }]);
+
   conn.release();
   pool[Symbol.dispose]();
+
+  it('should have received the query result', () => {
+    assert.deepStrictEqual(rows, [{ 1: 1 }]);
+  });
 
   it('should have closed the pool', () => {
     // @ts-expect-error: internal access
@@ -183,57 +145,12 @@ await describe('dispose should end the pool', async () => {
 
 await describe('dispose should handle end before dispose on pool', async () => {
   const pool = createPool({ connectionLimit: 1 });
+
   await pool.promise().end();
   pool[Symbol.dispose]();
 
   it('should have closed the pool', () => {
     // @ts-expect-error: internal access
     assert.strictEqual(pool._closed, true);
-  });
-});
-
-const clusterQuery = (cluster: PoolCluster, sql: string) =>
-  new Promise<RowDataPacket[]>((resolve, reject) => {
-    cluster.of('*').query<RowDataPacket[]>(sql, (err, rows) => {
-      if (err) return reject(err);
-      resolve(rows);
-    });
-  });
-
-await describe('PoolCluster should implement Symbol.dispose', async () => {
-  const cluster = createPoolCluster();
-  cluster.add('MASTER', getConfig());
-
-  it('should be a function', () => {
-    assert.strictEqual(typeof cluster[Symbol.dispose], 'function');
-  });
-
-  cluster[Symbol.dispose]();
-});
-
-await describe('dispose should end the pool cluster', async () => {
-  const cluster = createPoolCluster();
-  cluster.add('MASTER', getConfig());
-  const rows = await clusterQuery(cluster, 'SELECT 1');
-  assert.deepStrictEqual(rows, [{ 1: 1 }]);
-  cluster[Symbol.dispose]();
-
-  it('should have closed the pool cluster', () => {
-    // @ts-expect-error: internal access
-    assert.strictEqual(cluster._closed, true);
-  });
-});
-
-await describe('dispose should handle end before dispose on pool cluster', async () => {
-  const cluster = createPoolCluster();
-  cluster.add('MASTER', getConfig());
-  await new Promise<void>((resolve, reject) => {
-    cluster.end((err) => (err ? reject(err) : resolve()));
-  });
-  cluster[Symbol.dispose]();
-
-  it('should have closed the pool cluster', () => {
-    // @ts-expect-error: internal access
-    assert.strictEqual(cluster._closed, true);
   });
 });
